mvt_venv_requirements.txt
altair==5.5.0
attrs==25.3.0
bibtexparser==1.4.3
binaryornot==0.4.4
blinker==1.9.0
bracex==2.6
cachetools==6.2.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
filetype==1.2.0
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
monty==2025.3.3
mpmath==1.3.0
mrcfile==1.5.4
narwhals==2.3.0
networkx==3.5
numpy==2.3.2
orjson==3.11.3
packaging==25.0
palettable==3.3.3
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
pymatgen==2025.6.14
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.16.1
six==1.17.0
smmap==5.0.2
spglib==2.6.0
streamlit==1.49.1
streamlit-ace==0.1.1
streamlit-antd==0.8.3
streamlit-embeded==0.0.1
streamlit-file-browser==3.2.22
streamlit-image-select==0.6.0
streamlit-molstar==0.4.21
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.5.0
watchdog==6.0.0
wcmatch==10.1
